requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing==2.2.0
ipdb==0.13.13
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
matplotlib-inline==0.1.7
numpy==2.2.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
tzdata==2025.1
wcwidth==0.2.13
